# src/main/resources/swagger.yaml
openapi: 3.0.1
info:
  title: API de inglés
  description: La siguiente api muestra los endpoints a trabajar sobre el sistema
  version: 1.0.0
servers:
- url: https://linguish.up.railway.app
...
paths:
...
  /user/find/{id}:
    get:
      tags:
        - usuarios
      summary: Obtiene la información de un usuario por su id.
      parameters:
      - name: id
        in: path
        description: Identificador del cliente a obtener.
        required: true
        schema:
          type: string
      responses:
        200:
          description: (OK) Información del usuario obtenida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUserGet'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /user/find/all:
    get:
      tags:
        - usuarios
      summary: Busca todos los usuarios en la BD.
      responses:
        200:
          description: (OK) Información de los usuarios obtenida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUsersGet'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /user/save:
    post:
      tags:
        - usuarios
      summary: Inserta la información de un nuevo usuario a la BD.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyUserPost'
      responses:
        200:
          description: (OK) Usuario creado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUserPost'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /user/update/{id}:
    put:
      tags:
        - usuarios
      summary: Actualiza los datos de un usuario por su id.
      parameters:
      - name: id
        in: path
        description: Identificador del usuario a actualizar.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyUserPut'
      responses:
        200:
          description: (OK) Usuario actualizado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUserUpdate'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /user/delete/{id}:
    delete:
      tags:
        - usuarios
      summary: Elimina los datos de un usuario por su id.
      parameters:
      - name: id
        in: path
        description: Identificador del cliente a eliminar.
        required: true
        schema:
          type: string
      responses:
        200:
          description: (OK) Información del cliente obtenida correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessUserDeleted'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
          
#MODULE
...
components:
  responses:
    Unauthorized:
      description: (Unauthorized) no hay autorización para llamar al servicio
    NotFound:
      description: (NotFound) no se encontró la información
    BadRequest:
      description: (BadRequest) los datos enviados son incorrectos o hay datos obligatorios no enviados
    ServerError:
      description: Error en el servidor
  schemas:
...
    SuccessUserGet:
      type: object
      properties:
        username:
          type: string
          description: Nombre de usuario.
        email:
          type: string
          description: Email del usuario.
    SuccessUsersGet:
      type: array
      items:
        $ref: '#/components/schemas/SuccessUserGet'
    BodyUserPost:
      type: object
      properties:
        username:
          type: string
          description: Nombre de usuario.
        email:
          type: string
          description: Email del usuario.
        password:
          type: string
          description: Contraseña del usuario.
    SuccessUserPost:
      type: object
      properties:
        response:
          type: integer
          enum: [1]
          description: Bandera que indica si la llamada al servicio fue exitosa.
        idUsuario:
          type: integer
          enum: [222]
          description: Bandera que le corresponde al usuario.
    BodyUserPut:
      type: object
      properties:
        username:
          type: string
          description: Nombre de usuario.
        email:
          type: string
          description: Email del usuario.
        password:
          type: string
          description: Contraseña del usuario.
    SuccessUserUpdate:
      type: object
      properties:
        response:
          type: integer
          enum: [1]
          description: Bandera que indica si la actualización del usuario fue exitosa.
        idUsuario:
          type: integer
          enum: [222]
          description: ID del usuario actualizado.
    SuccessUserDeleted:
      type: object
      properties:
        response:
          type: integer
          enum: [1]
          description: Bandera que indica si la eliminación del usuario fue exitosa.
        idUsuario:
          type: integer
          enum: [222]
          description: ID del usuario eliminado.
          
#MODULE
...